+++ examples/sveltekit-magic-link/src/lib/types.ts
@@ -1,0 +1,4 @@
+export interface TestTable {
+	id: string;
+	created_at: string;
+}