+++ packages/nextjs/src/utils/log.ts
@@ -1,6 +1,21 @@
-import * as loglevel from 'loglevel';
+const dev = process.env.NODE_ENV !== 'production';
 
-const logger = loglevel.getLogger('@supabase/auth-helpers-nextjs');
-logger.info('Starting @supabase/auth-helpers-nextjs logging...');
+const logger = {
+  log: (message?: any, ...optionalParams: any[]) => {
+    dev ? console.log(message, ...optionalParams) : null;
+  },
+  error: (message?: any, ...optionalParams: any[]) => {
+    console.error(message, ...optionalParams);
+  },
+  info: (message?: any, ...optionalParams: any[]) => {
+    logger.log(message, ...optionalParams);
+  },
+  debug: (message?: any, ...optionalParams: any[]) => {
+    logger.log(message, ...optionalParams);
+  },
+  warn: (message?: any, ...optionalParams: any[]) => {
+    dev ? logger.error(message, ...optionalParams) : null;
+  }
+};
 
 export default logger;