+++ packages/sveltekit/src/handlers/index.ts
@@ -1,3 +1,4 @@
+export { handleSession } from './session';
 export { handleCallback } from './callback';
 export { handleUser } from './user';
 export { handleLogout } from './logout';