+++ packages/sveltekit/src/instance.ts
@@ -3,7 +3,7 @@ import { getClientWithEnvCheck } from './utils/initSupabase';
 let supabaseUrl: string | undefined;
 let supabaseAnonKey: string | undefined;
 
-export const skHelper = (url?: string, anonKey?: string) => {
+export const createSupabaseClient = (url?: string, anonKey?: string) => {
   if (!supabaseUrl || !supabaseAnonKey) {
     supabaseUrl = url;
     supabaseAnonKey = anonKey;
@@ -18,3 +18,8 @@ export const skHelper = (url?: string, anonKey?: string) => {
     })()
   };
 };
+
+/**
+ * @deprecated use `createSupabaseClient` method instead
+ */
+export const skHelper = createSupabaseClient;