+++ packages/sveltekit/src/utils/getUser.ts
@@ -1,6 +1,6 @@
 import { createClient, type ApiError } from '@supabase/supabase-js';
 import type { User } from '@supabase/supabase-js';
-import { skHelper } from '../instance';
+import { createSupabaseClient } from '../instance';
 import {
   setCookies,
   parseCookie,
@@ -50,7 +50,7 @@ export async function getUser(
   try {
     const {
       apiInfo: { supabaseUrl, supabaseAnonKey }
-    } = skHelper();
+    } = createSupabaseClient();
     if (!supabaseUrl || !supabaseAnonKey) {
       throw new Error(
         'supabaseUrl and supabaseAnonKey env variables are required!'
@@ -121,11 +121,8 @@ export async function getUser(
     if (e instanceof JWTPayloadFailed) {
       logger.info('JWTPayloadFailed error has happened!');
       response.error = e.toObj();
-    } else if (e instanceof CookieNotFound) {
-      logger.warn(e.toString());
     } else if (e instanceof AuthHelperError) {
-      logger.info('AuthHelperError error has happened!');
-      logger.error(e.toString());
+      // do nothing, these are all just to disrupt the control flow
     } else {
       const error = e as ApiError;
       logger.error(error.message);