+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @supabase/auth-helpers-react
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [56228e3]
+  - @supabase/auth-helpers-shared@0.1.3
+
 ## 0.2.2
 
 ### Patch Changes